requirements.txt
admin==0.0.1
alembic==1.7.7
Babel==2.10.1
blinker==1.4
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.2
clikit==0.6.2
crashtest==0.3.1
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
filelock==3.6.0
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Sessionstore==0.4.5
Flask-SQLAlchemy==2.5.1
Flask-Toastr==0.5.7
Flask-WTF==1.0.1
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
keyring==23.5.0
lockfile==0.12.2
Mako==1.2.0
MarkupSafe==2.1.1
msgpack==1.0.3
packaging==20.9
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pylev==1.4.0
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
secondary==1.0
sentry-sdk==1.5.9
shellingham==1.4.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.35
success==1.0.0
tomlkit==0.10.1
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
